+++ legatium-webclient-logging/src/test/kotlin/eu/inqudium/legatium/webclient/logging/ClientRequestLoggingFilterIntegrationTest.kt
@@ -35,8 +35,8 @@ import java.util.concurrent.TimeoutException
     classes = [IntegrationApp::class],
     webEnvironment = SpringBootTest.WebEnvironment.NONE,
     properties = [
-        "client-logging.log-request-body=true",
-        "client-logging.log-response-body=true",
+        "client-logging.log-request-body=always",
+        "client-logging.log-response-body=always",
         "client-logging.response-headers.includes=Content-Type",
         "client-logging.response-headers.unmasked=Content-Type",
         "client-logging.request-headers.includes=X-Correlation-Id",